Suppress UUID warning in map_instance unit tests
nova-manage cells_v2 map_instances call uses a non canonical UUID
serialization to store the instance marker in the DB. The
oslo.versionedobjects UUIDField emits a warning. A later patch would
like to turn this warning to an error during the unit and functional
test to avoid adding new violations.

As the underlying DB schema is not violated this patch proposes to
suppress the warning in the affected unit tests.

Change-Id: I5b11b9df26e4e38516b5674e0e6c1fc79527129b