Make nova service-list use scatter-gather routine
This patch makes nova service-list use the scatter-gather routine
so that if a cell is down, at least the services from other cells
are listed by ignoring the down cell instead of the whole command
failing with an API exception as is the current situation. Also
making this query parallel for all cells is more efficient.

Depends-On: https://review.openstack.org/569112/

Change-Id: I90b488102eb265d971cade29892279a22d3b5273
Closes-Bug: #1726310