Add requested_resources field to RequestSpec
The new RequestSpec.resource_requests stores a list of RequestGroup
objects each representing a granular resource request group requesting
resource for the given instance.

This will be used in a later patch to communicate the Neutron port's
resource request to the scheduler. Also this can be used later to
refactor flavor-based resource request handling.

This new field is intentionally not persisted to the DB for two
reasons:

* the port's resource request is already persisted in neutron
  so this would be a data duplication that could lead to data
  inconsistencies
* this field is only used to carry information from the the
  nova-conductor to the nova-scheduler for each select_destinations()
  request

blueprint: bandwidth-resource-provider
Change-Id: I53e5debcffd6de2b3a2ff838e7f5da33fa1a82b8