Enforce placement minimum in nova.cmd.status
We keep forgetting to bump the minimum required placement version in
nova.cmd.status (and all the related bits and pieces) whenever we
change the report client to require a new version.

This patch interposes a check in the test_report_client functional suite
any time get/put/post/delete is called from the report client.  If we
see a microversion higher than the minumum specified in nova.cmd.status,
we raise an exception, which will blow up the test.  This should force
the author of a new patch on SchedulerReportClient to do the necessary
paperwork in that patch.

...assuming said author happens to write a test in test_report_client.
This pattern can and should be copied into other test suites where
report client tests are likely to be written, to broaden the scope of
this enforcement.

Change-Id: I5482b92f941261ab6ee6b7cd532ce268c31fe793