ProviderTree.has_inventory_changed for new fields
The has_inventory_changed method of ProviderTree is deliberately not
automatically reporting True if the incoming record is "sparse" - that
is, for a given resource class, it only contains a subset of the keys in
the existing record.  For example, the caller may want to know if the
total has changed, but doesn't care about the other fields, so may ask:

ptree.has_inventory_changed(uuid, {'VCPU': {'total': 3}, ...})

...and has_inventory_changed will only compare the 'total' field.

However, we were doing the same if the incoming record contained fields
that were *not* in the existing record.  For example, if the current
inventory was:

{ 'VCPU': {
      'total': 8,
      'min_unit': 1,
      'max_unit': 8,
      'step_size': 1,
  }
}

...and the new inventory was:

{ 'VCPU': {
      'total': 8,
      'min_unit': 1,
      'max_unit': 8,
      'step_size': 1,
      'allocation_ratio': 16.0,  # <== new
      'reserved': 1,             # <== new
  }
}

...the presence of the new fields was *not* enough to trigger a True
response.

This is problematic because we currently expect to be able to add
specifically these fields - allocation_ratio and reserved - to an
inventory record if they are not present (see [1]).

With this change set, we add a condition to automatically return True
(inventory *has* changed) whenever the incoming (new) record contains
fields absent from the existing record.

[1] https://github.com/openstack/nova/blob/050e9202f69b6eeb23026b7350eeb8821ca9f1cc/nova/compute/resource_tracker.py#L83-L125

Partial-Bug: #1771728
Change-Id: I80dd6d08322bcedc8535f04baf2e96473f5e3aeb