libvirt: Do not reraise DiskNotFound exceptions during resize
When an instance has VERIFY_RESIZE status, the instance disk on the
source compute host has moved to <instance_path>/<instance_uuid>_resize
folder, which leads to disk not found errors if the update available
resource periodic task on the source compute runs before resize is
actually confirmed.

Icec2769bf42455853cbe686fb30fda73df791b25 almost fixed this issue but it
will only set reraise to False when task_state is not None, that isn't
the case when an instance is resized but resize is not yet confirmed.
This patch adds a condition based on vm_state to ensure we don't
reraise DiskNotFound exceptions while resize is not confirmed.

Closes-Bug: 1774249
Co-Authored-By: Vladyslav Drok <vdrok@mirantis.com>
Change-Id: Id687e11e235fd6c2f99bb647184310dfdce9a08d