Use list instead of set for duplicate check
There can be legitimately distinct allocation requests with the same
combination of providers. For example, in a case where where both
SHR1 and SHR2 have both CUSTOM_RC1 and CUSTOM_RC2, the following two
allocation requests are different:

 AllocationRequest(anchor=CN, resource_requests=[
     AllocationRequestResource(rp=CN, resources=[VCPU: 1]),
     AllocationRequestResource(rp=SHR1, resources=[CUSTOM_RC1: 1]),
     AllocationRequestResource(rp=SHR2, resources=[CUSTOM_RC2: 1]),
 ])

 AllocationRequest(anchor=CN, resource_requests=[
     AllocationRequestResource(rp=CN, resources=[VCPU: 1]),
     AllocationRequestResource(rp=SHR2, resources=[CUSTOM_RC1: 1]),
     AllocationRequestResource(rp=SHR1, resources=[CUSTOM_RC2: 1]),
 ])

But getting allocation candidates, placement had been filtering out
one of the two assuming them as the same allocation requests because
the combination of the providers in play was the same.

This was because placement filtered duplicate candidates using a set
of provider ids in play while we can't consider resource information
with the set. This patch fixes the bug by using the list instead of
the set. This is because we can put resource information in the order
of that list. Namely, in the example above, we can't detect
"set([CN, SHR1, SHR2])" is different as "set([CN, SHR2, SHR1])", but
we can detect [CN, SHR1, SHR2] is different from [CN, SHR2, SHR1].

Change-Id: I90627dfe85661118c4ec2571fc32ffca99dd73d2
Blueprint: nested-resource-providers-allocation-candidates
Closes-Bug: #1772243