Change the ServerGroupAntiAffinityFilter to adapt to new policy
In this patch, the anti-affinity filter will get the
max_server_per_host limit from policy rules, and compare it
against the number of servers within the same group on a given
host. If the filter finds the number is not satisfied with the
limit, it will filter out this host.

The default ``max_server_per_host`` for the anti-affinity filter
is 1 for backward compatibility.

Change-Id: I4b67ec9dd4ce846a704d0f75ad64c41e693de0fb
blueprint: complex-anti-affinity-policies