Add functional test for AggregateMultiTenancyIsolation + migrate
A bug was reported against Ocata where a non-admin user
creates a server and the user's project is isolated to a
set of hosts via the AggregateMultiTenancyIsolation filter.

The admin, with a different project, cold migrates the server
and the filter rejects the request because before change
I195d389ac59574724a5e7202ba1a17d92c53a676 the cold migrate
task would re-generate the RequestSpec using the request context
which was from the admin, not the owner of the instance.

Even though this is not a problem past Ocata, we did not have
functional test coverage for this scenario so it is added here.

This will also be used to backport the fix to Ocata to show
the regression and fix in that branch.

Change-Id: I97559607fc720fb98c3543ff3dd6095281752cd4
Related-Bug: #1774205
Related-Bug: #1675607