Adapt _validate_instance_group_policy to new policy model
To solve the race condition problem, we do anti-affinity policy
sanity check in the instance build code of compute manager, but
this check is only valid in original policy.

Specifically, this change adds validation of the new
max_server_per_host policy rule which can be used to allow >1
anti-affinity group member on the same host.

blueprint: complex-anti-affinity-policies

Change-Id: I755b6fdddc9d754326cd9c81b6880581641f73e8