Mention running rootwrap in daemon mode if hitting vif plug timeouts
Per an operators mailing list thread [1] people can hit vif
plug timeouts at scale and a possible solution is running
rootwrap in daemon mode in the neutron agent, which several
operators have said resolves the issue at scale for them.

This change adds a mention of that configuration in the
"vif_plugging_timeout" configuration option since that
might be the first place operators look when they start
hitting this issue.

A corresponding change will be made to neutron to improve
the config option help text for the "root_helper_daemon"
option.

[1] http://lists.openstack.org/pipermail/openstack-operators/2018-May/015364.html

Change-Id: I65c0d518bc1c1a0d94fa7ac01294b03cce7d3752
Closes-Bug: #1774217