ironic: bugfix: ensure a host is set for volume connectors
Somehow, in some not fully understood case,
we have discovered that we are not storing a
host field for the volume connector information.

This information is largely required for logging,
and in some cases back-end driver relationships,
but previously we would only set the field in the
case that we have an IP address. That is not ideal
in that some users may be using other transport
mediums like fibre channel.

Change-Id: I17b1aaa26b7e7aff3ebb73fa10ab576149a00a3f
Related-Bug: #1766661