Fix bug to filter_scheduler
In https://docs.openstack.org/nova/latest/configuration/config.html#filter_scheduler.max_instances_per_host
refers to the "num_instances_filter" but that's the name
of the python module, not the actual filter.
So the config option help text should be updated to use
the actual filter names that would go in the 'enabled_filters'
option, not the python module in which those filters live.

Change-Id: Icf7e64ed45fa5013714d85ea22794f106d4e387c
Closes-bug: #1774676