[placement] Do not import oslo_service for log_options
Placement does not use oslo_service, but it is imported for one single
configuration option: "log_options".

This change duplicates that option info nova/conf/placement.py (in the
DEFAULT group) so that oslo_service doesn't need to be imported and its
dependencies (e.g., eventlet) are also not imported.

This is relevant for at least two reasons:

* When running placement in isolation (such as a container) we don't
  want to have to include packages aren't really used

* When placement is extracted we don't want to require things (in
  requirements.txt) that aren't really used.

Change-Id: I5a19cd3e18bb8c0e0ad2d79f54194f8b941ce307