Move db using provider unit tests to functional
Some of the ResourceProvider unit tests used the db directly.

These have been moved to the functional test_resource_provider.py and
transformed as required to make them work in that setting. Some test
cases were appended to existing tests where it made sense.

The save_immutable test which claimed to use the db, but did not,
remains in the original file but in a more suitable test class.

Change-Id: I7a91ed797b052500fdd80690cb35f1049ef6c422