objects: Add RequestSpec.network_metadata
We need a way to pass information about networks from the conductor to
the scheduler. These are not really part of the network requests but
rather metadata about these networks, meaning the 'NetworkRequest'
object doesn't really make sense. Instead, we use the recently added
'NetworkMetadata' object like the good citizens we are.

Part of blueprint numa-aware-vswitches

Change-Id: Icb295bbd8c83e2e340a7ac3ecc1f159e0db7c7b1