Keep attach_mode as top-level field in _translate_attachment_ref
The attach_mode is a top-level field on the attachment record in Cinder.
There is an access_mode entry in the connection_info field which is used
by the volume driver in compute when making an attachment, but the
attach_mode field is read by the API during swap volume so we don't have
to dig into the connection_info details.

This changes _translate_attachment_ref to leave attach_mode on the
top-level attachment reference dict so the API does not need to look
into the connection_info data.

Change-Id: I1207687d1b00416482ac34ad3250e7121d9b4352