Make check_can_live_migrate_destination use long_rpc_timeout
This call results in several other RPC calls, which makes it a prime
target for a cascading failure if things are moving slowly. Move this
to the long_rpc_timeout bucket where it can use heartbeats to let the
call go longer if needed.

There was no test_rpcapi test for this method at all, so this adds
one and tweaks the helper as needed to work with it.

Change-Id: I2597ea46efce7c27db4745685f9fabb9da9a883e