Update the disk_cachemodes to mention an rbd detail
The disk_cachemodes text for the writeback mode explains the danger
inherent with providing writeback cache semantics to the guest where
the host page cache is involved. The qemu-native RBD driver provides
correct flush behavior to the RBD cache on the host, and thus is
safe to use (and recommended by the upstream Ceph documentation).

http://docs.ceph.com/docs/master/rbd/qemu-rbd/#qemu-cache-options

Change-Id: I975a6544484bc8d5f062e99f85374710295443c7