Remove mox in unit/network/test_neutronv2.py (11)
Replace mox with mock in the following methods
in the TestNeutronv2 class
in nova/tests/unit/network/test_neutronv2.py.
Then moves the methods to the TestNeutronv2WithMock class.

* test_validate_networks_ex_1
* test_validate_networks_ex_2
* test_validate_networks_duplicate_enable
* test_validate_networks_not_specified
* test_validate_networks_port_not_found
* test_validate_networks_port_show_raises_non404
* test_validate_networks_port_in_use

The 'test_validate_networks_ex_1' test case was broken.
The expected behavior of the 'validate_networks' method is raising
the NetworkNotFound exception, but the exception was not raised.
so fix it.

Change-Id: Ifeac24ea65065575c2aaf7946835f70cdb763012
Implements: blueprint mox-removal-train