Fix regression when listing build_requests with marker and ip filter
Change Ic02206e887e3fea7752d615bbed680510c482097 attempted
to optimize the GET /servers flow by skipping filtering on
build requests if the ip or ip6 filters were used, since
servers that are not yet scheduled (build requests) can't have
ips. However, if a marker is provided and the marker is in the
build_requests table, we fail to look there and then check the
cells for the marker, which won't exist and result in a 400
MarkerNotFound error.

This fixes the issue by *not* skipping build requests if there
is a marker specified. A functional test is added which will
show the 400 MarkerNotFound error if the code fix is removed.

Change-Id: Ibdd157d06252c3c0219539ec798c8d9d3a8ae26f
Closes-Bug: #1777458