Remove mox in unit/network/test_neutronv2.py (22)
Replace mox with mock in the following methods
in the TestNeutronv2 class
in nova/tests/unit/network/test_neutronv2.py
and moves the methods to the TestNeutronv2WithMock class.

* test_get_physnet_tunneled_info_multi_segment
* test_get_physnet_tunneled_info_vlan_with_multi_segment_ext
* test_get_physnet_tunneled_info_multi_segment_no_physnet
* test_get_physnet_tunneled_info_tunneled
* test_get_phynet_tunneled_info_non_tunneled
* _test_get_port_vnic_info
* test_get_port_vnic_info_1
* test_get_port_vnic_info_2
* test_get_port_vnic_info_3
* test_get_port_vnic_info_requested_resources

Then the TestNeutronv2 class is removed and
the TestNeutronv2WithMock class is renamed to 'TestNeutronv2'.
Comments about replacing mox are removed.

This patch completes replacing mox with mock in the file.

Change-Id: I57040ea4457e53ad97d536591fd31808a5a6c227
Implements: blueprint mox-removal-ussuri