Ironic update_provider_tree: restore traits override
The preceding change [1] converted get_inventory+get_traits to
update_provider_tree in the ironic virt driver.  It used
provider_tree.add_traits to set the traits on the compute node provider.
This was a deviation from the previous behavior, which used
update_traits (in the resource tracker) thereby *replacing* the compute
node provider's traits with whatever the driver reported.  This has the
drawback of obliterating any traits set outside of the compute service,
but that's a bridge we'll cross later.

This patch changes add_traits to update_traits to restore the previous
behavior.

[1] https://review.openstack.org/#/c/576588/

Change-Id: Ic9729a02b502c89570d967e27bc169f8b2036488