[placement] Extract create_allocation_list
The data assembly in set_allocations which results in an
AllocationList will be useful for forthcoming /reshaper code, so
here we extract it to its own public module-level method.

As noted in the associated TODO, there's some refactoring that can
be done to get methods like this in a better place while keeping the
handlers concise. util.py is getting too big and too diverse, so
there is not ideal.

Change-Id: I4b69ddaa6dbb66f977c5dad2d508cac3192d19d4